Infrastructure/src/main/java/org/cris6h16/Config/SpringBoot/Security/UserDetails/UserDetailsWithId.java
package org.cris6h16.Config.SpringBoot.Security.UserDetails;

import org.springframework.security.core.GrantedAuthority;
import org.springframework.security.core.userdetails.User;

import java.util.Collection;

// not necessary with jwt
public class UserDetailsWithId extends User {
    private final Long id;

    public UserDetailsWithId(String username, String password, boolean enabled, boolean accountNonExpired, boolean credentialsNonExpired, boolean accountNonLocked, Collection<? extends GrantedAuthority> authorities, Long id) {
        super(username, password, enabled, accountNonExpired, credentialsNonExpired, accountNonLocked, authorities);
        this.id = id;
    }

    public Long getId() {
        return id;
    }
}